add component SessionLogoutTimeout, add sound
