+++ frontend/src/locales/de.json
@@ -199,7 +199,8 @@
   "send_per_link": "GDD versenden per Link",
   "session": {
     "warningText": "Die Session läuft gleich ab!",
-    "lightText": "Du wirst automatisch abgemeldet in"
+    "lightText": "Du wirst automatisch abgemeldet in",
+    "extend":"Verlängern"
   },
   "settings": {
     "language": {
