+++ frontend/src/locales/en.json
@@ -199,7 +199,8 @@
   "send_per_link": "GDD send via link",
   "session": {
     "warningText": "The session is about to expire!",
-    "lightText": "You will be automatically logged out in"
+    "lightText": "You will be automatically logged out in",
+    "extend":"Extend"
   },
   "settings": {
     "language": {
